Fix cross-fade crash using wxMemoryDC::SelectObjectAsSource
The problem: Calling wxMemoryDC::SelectObject will attempt to turn a
shared wxBitmap into a non-shared wxBitmap using wxBitmap::UnShare.  But
UnShare calls wxBitmap::CopyFromDIB, which does not correctly copy the
underlying wxBitmapRefData.

This scrambles the wxBitmap sufficiently that we can no longer use raw
access, which in turn causes the cross-fader to crash with an assertion
failure.

For the whole bogus, raving saga, see:

  http://groups.google.com/group/wx-users/browse_thread/thread/2a0e1d2aa4ed364e

Our strategy:

 1) Work around the immediate problem by using SelectObjectAsSource
    whenever we need a read-only wxMemoryDC.  This _appears_ to make
    Halyard work, although it's not a general solution.
 2) Notify the wxWidgets developers of the underlying problem (see the
    above link).
 3) If the wxWidgets developers confirm that wxBitmap is supposed to
    have copy-on-write semantics, we'll need to replace wxBitmap with
    shared_ptr<wxBitmap> in many places in Halyard, if only to prevent
    accidental copies from killing performance.