ScriptEditor: Always use standard syntax highlighting
Historically, we've relied on a patched wxScintilla LexLisp.cpp to get
our syntax-highlighting right.  When we ported the ScriptEditor to run
on wxWidgets 2.8 on the Mac, Brian was able to eliminate most of this
custom code.

Here, we switch to using the wx 2.8/2.9 code all the time, and add a
patch to wxScintilla highlighting keyword argument names using
wxSTC_LISP_SYMBOL.  We also reorder our table of syntax highlighting
colors to put everything back in the proper order.

Eric Kidd (1):
      stc: Syntax-highlight Lisp keyword arguments as symbols