ScriptEditor: Properly handle enter key in find dialog
Because wxTE_MULTILINE and wxTE_PROCESS_ENTER are not really compatible,
it's fairly tricky to intercept the Enter key and use it to dismiss the
dialog.  The only way to do this is to override wxEVT_KEY_DOWN on the
wxTextCtrl objects.  Overriding wxEVT_CHAR fails, because it occurs
after the Enter key has already been added to the wxTextCtrl's value.

With this patch, the ScriptEditor appears to be working correctly
again.