DrawingArea: Add basic Cairo support
DrawingArea has long relied on a mix of wxDC calls and custom drawing
routines.  But neither approach is really satisfactory: wxDC calls can't
handle alpha channels acceptably well, and custom drawing routines are
time-consuming to write, so we've never had very many.

This patch adds Cairo support to DrawingArea, allowing us to access a
rich set of graphics primitives.

Note that on Windows, we're currently linking against a stripped-down
build of Cairo, with no PNG or typography support, and only a handful
of different backends.