DrawingArea: Add support for DRAW-OVAL and DRAW-OVAL-OUTLINE
We use Cairo to implement support for ovals.  This is highly similar to
our support for rectangles--it should be possible to draw an oval on top
of a rectangle, and have the sides of the rectangle appear as tangents
to the oval's curve.

There's still some code duplication here which we need to remove, but I
figured it was best to get the working Cairo code committed before
trying to refactor it.