DrawingArea: Draw letters using Cairo
This required a fairly major overhaul, because our PixMap class didn't
use a memory layout compatible with Cairo.  We had two choices: (1) Copy
the data out of the PixMap into a local buffer, or (2) modify the call
to DrawPixMap so that we receive appropriately-formatted data in the
first place.  I went with choice (2).

What's changed:

 - The PixelMap<> template and the PixMap class are now gone, along with
   quite a bit of supporting machinery.
 - The GreyMap class now stands alone, and is guaranteed to use a Cairo-
   compatible memory layout.
 - DrawingArea now implements DrawGreyMap instead of DrawPixMap.

We could probably simplify the code even further (and perhaps speed it
up) by allowing Cairo to manage our glyph cache.  But that's a project
for another day.