kernel.ss: Cache results of have-prim?
We are currently calling the WakeUpIfNecessary primitive each time
(call-prim ...) is called.  We also call (has-prim? ...) to determine
if WakeUpIfNecessary is available.  This patch adds a hash table to
has-prim? to cache the results, improving our primitive call time
substantially.  We get a 10-20% performance increase on creating a
large number of elements by applying this patch.