nodes.ss: Cache node .full-name computation
We call .full-name on every node before calling any primitive that
manipulates the underlying Zone or Overlay.  This is a recursive
process, that computes the full name of the parent, and appends
a / and the node's name.  As this is fairly expensive, and called
for a large number of primitive calls, this patch adds caching of
a node's name in a slot.

Performance measurements were a bit variable for this patch, but
it looks like it buys us about 10% performance improvement on
creating elements.

Paired-with: Eric Kidd <eric.m.kidd@dartmouth.edu>