BREAKS MAC: DrawingArea: Use a cairo_image_surface instead of a wxBitmap
We were using a wxBitmap to hold the data associated with our
DrawingArea, but this didn't work on Windows, because there's no way to
treat a wxBitmap with an alpha channel as a Cairo surface.

This patch rips out the wxBitmap code, and replaces it with equivalent
cairo_image_surface code.  We need to use a Cairo image surface (and not
an accelerated, platform-specific offscreen surface) because we need to
call cairo_image_surface_get_data in our MaybeInitializeGameOverlay
and GetPixel functions.  (We use the latter for hit-testing overlays.)

This also requires us to convert a few other functions here and there
which try to treat a DrawingArea as though it were a wxBitmap or wxDC.

This patch fixes about half the drawing problems with the Windows
engine, but it breaks the Mac engine.

Modified submodules:

* libs/quake2 6a1fe47...26b0ecc (1):
  > Overlays: Add Q2_FORMAT_BGRX for use with Cairo