Bug fix: Initialize :shown? on %custom-element% correctly
In commit 74031166 (Performance: Don't set element parameters to default
values on setup) we added .FINISH-INITIALIZING-ENGINE-NODE to set up
parameters on nodes that are not set up upon creation.  We forgot
to call (super) when we overrode this on %custom-element%, so we did
not inherit the setting of .SHOWN? in the engine.  This patch fixes
that mistake, and adds a test card to help prevent future regressions.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?17014