Performance: Remove extra primitive call to WakeUpIfNecessary
When a wait on media ends, we need to tell the kernel to wake up.  We
do this by setting a flag that is checked in %kernel-check-state (which
is called in various contexts, including every CALL-PRIM).  This flag
had been handled by Stage in C++, and checking it and waking up handled
by a primitive call WakeUpIfNecessary.  This introduced an extra prim
call for (almost) every prim call.  This patch moves the state and logic
into kernel.ss.

This patch improves performance for untraced primitives by about 50%;
that is, they take about half the time, and for traced primitives by
about a third.  This translates to an improvement of about 15% in
element creation/deletion.