Performance: Only call StateDbUnregisterListeners if we have listeners
This patch removes an extra primitive call to
StateDbUnregisterListeners for elements that have never had any state
db listeners registered.  We set a slot on a node whenever we register
a listener, and check that slot before unregistering listeners.  To do
this, we moved the unregister listeners logic from kernel.ss into
state-db.ss, as a monkey-patched ADVISE BEFORE.

This improves performance on element creation about 15-20% (about 1.8ms
to 1.5ms).