0.5.33 - 4 June 2009
This release introduces a few more performance optimizations.  At this
point, we believe that we are doing no unnecessary primitive calls for
basic element creation, and that primitive call overhead is no longer
the bottleneck.  Since release 0.5.30.1, we have improved performance on
element creation by a factor of around 3, and on primitive call overhead
by a factor of around 20.

This release updates draw-text and draw-graphic to return the bounding
box of the item drawn.  This can help DRAW methods avoid an extra
primitive call to measure text or a graphic.

Brian Campbell (3):
      Performance: Remove extra primitive call to WakeUpIfNecessary
      Performance: Only call StateDbUnregisterListeners if we have listeners
      Performance: Disable trace logging for primitives

Eric Kidd (4):
      Bug fix: Only expand WITH-DC argument once
      Performance: Return bounding box from draw-text and draw-graphic
      Don't build log messages for primitives if we don't need them
      Merge branch 'master' of ssh://imlsrc.dartmouth.edu/var/lib/git/halyard