StartupDlg: Implement "recent files" list
This dialog has been sitting around, completely useless, for quite a few
years.  Let's do something with it.

Note that we currently update wxFileHistory from the GUI code, and not
the lower-level code, because we only want to record files in response
to real user actions.

Note that if the copy of Halyard appears to be part of an existing
program, we also "pre-load" the recent files list with that program.