wxAui: Convert StageFrame to use wxAui
wxAui is a high-level user-interface library for wxWidgets.  It attempts
to recreate much of the functionality of UIs such as Visual Studio,
including dockable palettes, configurable toolbars, and a center
"document area" within a frame.

This patch converts StageFrame to use wxAuiManager instead of our
exisitng SashFrame class.  Some highlights of this patch include:

 - Following standard wxAui terminology, we now refer to saved frame
   configurations as "perspectives", not "layouts".
 - We now require wxWidgets 2.8.8 or later, because we need access to
   recently-added wxSizer functions.  I've set configure.ac to look
   for wxWidgets 2.8.10, because that's what I've tested against.
 - We can no longer set a STAGE_FRAME_COLOR without messing up wxAui's
   drawing methods.  So instead, I've tried experimentally setting
   wxCLIP_CHILDREN on StageFrame, though this may not actually be
   necessary these days.
 - ProgramTree is now a subclass of wxWindow.  We may (or may not) want
   to collapse ProgramTree into wxTreeCtrl object that it contains for
   the sake of simplicity.

The real meat of this patch, however, appears in three functions:

  StageFrame::Layout
  StageFrame::UpdateMinimumFrameSize
  StageFrame::ShowFullScreen

The code in UpdateMinimumFrameSize, in particular, is fairly tricky--it
was arrived at after several hours of trial and error, and it appears to
work on wxMac 2.8.10 and wxMSW 2.9.  But without this code, we can't
correctly maintain minimum frame sizes, and we get lots of ugly
overlapping panes.

Modified submodules:

* libs/wxWidgets 08327cc...ed0183f (1):
  > Use non-DLL versions of C runtime for wxAUI