Remove ancient timecoder demo
This code was added when wx5L (the ancestor of Halyard) was only a
couple of weeks old.  It never actually did anything useful, because it
was hardcoded to a specific QuickTime clip.  We need to remove this
before we can get rid of our ToolWindow class, after which we can merge
the remaining ToolWindow subclass (Listener) into our new wxAui GUI.