ProgramTree: Get rid of useless "Cards" node
Our program tree used to contain an extra "Cards" node:

  Program 'Foo'
    Cards
      start
      card2
      card3

This patch combines the functions of the "Program" node and the "Cards"
node, yielding the following structure:

  Program 'Foo'
    start
    card2
    card3

This was slightly tricky, because the "Program" node now represents both
a model::View of the file application.halyard, and also the root node in
our script.  When we actually implement the GUI editor, we'll need to
permanently fuse these two concepts.