wxAui: Allow user to reset frame perspective to default
When debugging the wxAui code, I often need to clear the saved
perspective from the preference file.  This is especially necessary when
adding new panes, because they will always be hidden when loading a
saved perspective.

This patch moves a bunch of setup and teardown code in AuiFrame, and
adds the necessary code to restore the initial perspective.