Cleanup: Modify fix-comments to strip "/////" clutter
This script now converts comments of the form:

  //////////
  /// Foo.
  ///
  /// Bar.
  ///

To:

  /// Foo.
  ///
  /// Bar.

We've been using the new form for a while, but there was still quite a bit
of code using the old style.  Note that we only fix these comments if
they're indented, because the top-level ones are rarer and less
offensive, and some of them may be serving semi-useful purposes.