Cairo: Remove warning for drawing 0-area rectangle outline
We have a warning for the case of drawing a rectangle with an outline
too thick to fit in the overlay it is being drawn in.  This warning
was causing many warnings in SAE, where some elements with 0 area had
a rectangle drawn in them.  Because this case is relatively simple
and we already have code preventing them from doing anything bad, we
can relax our check, to only give a warning if we have a non-0 width
or height and stroke width too large to fit.

Paired-with: Eric Kidd <eric.m.kidd@dartmouth.edu>