Cleanup: Add tool for updating brace style
This is a fairly ugly and simple-minded piece of code, but it gets the
job done.  See the comment at the top of the file for details on how
this works.

Before writing this script, I tried using a wide variety of code
formatters, but they all insistent on reindenting code (badly) when fixing
the braces, and our indentation is fine.