Cleanup: Convert braces to Java style
The traditional K&R brace style is standard for C:

  void k_and_r()
  {
      if (cond) {
          bar();
      } else {
          baz();
      }
  }

Most of our code over the last several years uses a Java style, which is
highly similar, but more compact if you have lots of small functions:

  void java() {
      if (cond) {
          bar();
      } else {
          baz();
      }
  }

Our older code tends to put each brace on its own line:

  void old_style()
  {
      if (cond)
      {
          bar();
      }
      else
      {
          baz();
      }
  }

This style uses quite a bit of vertical whitespace, which makes it
inconvenient on a laptop screen.  (Especially since I tend to use
larger fonts for code.)

We've grown tired of inconsistent brace placement and wasted vertical
whitespace over the last several years.  As part of our
once-in-a-blue-moon code reformatting pass, we want to convert all code
to consistently use the new style.

Note that a few of the newly converted lines may exceed 80 columns.  These
will be fixed when we encounter them in the source.

Also note that trailing whitespace has been removed from a few of the lines
that we did in-depth processing on before deciding not to move a brace
character.