/m4: Pick correct mzscheme libs based on $ac_mzscheme_want_precise_gc
This patch fixes an error in the mzscheme m4 file whereby the precise
gc libs were being used when $ac_mzscheme_want_precise_gc was "0" and
the imprecise gc libs were being used when the value was "1".

For the record, "cgc" is the imprecise gc and "3m" is the precise gc.

Paired-with: Brian Campbell <brian.p.campbell@dartmouth.edu>