gnome-autogen.sh: Look for glibtool on the Mac
We don't actually need libtool yet, but Robinson is experimenting with
an Ubuntu port of Halyard.  In case he needs libtool, I'm going to
update gnome-autogen.sh to know about glibtool on the Mac.  Note that
MacOS X 10.5 includes glibtool 1.5.26, which is moderately old and
doesn't support some of the newer macros.

This code is vaguely inspired by a number of glibtool workarounds that
were found using Google, but doesn't particularly resemble any of them
closely, because we're using a very new version of gnome-autogen.sh.