ASSERT: Check condition at call site, not in separate function
Up until now, we've been passing the argument of ASSERT to another
function, which checks whether or not the condition is true.  This is
problem, because it's hard to pass a pseudo-boolean 'if' condition as
a function parameter.  See the most recent part of this saga in the
following patches:

  9d6aed2aed5ae54156c6cbb5d35d9f9219b63529
  a3ab6597a467a9bdc33f4da52a2d5cac696cad59
  a54dd201627777bfbad0bee897ea8dbe7693c3bd

As Taylor Campbell points out, we'd technically need to use intptr_t,
which isn't generally available, and which _still_ doesn't work with
boost::shared_ptr.

We could use "!!(test)" to coerce the test to a conditional.  But
really, we're probably better off using an inline 'if' statement inside
our ASSERT macro.