C++ nodes: Modify Wait primitive to look up element like other primitives
As part of the refactoring to create a C++ Node class, we want to
eliminate as many references to mElements and its iterators as possible.
In this case, the Wait primitive used a non-standard way of looking up
what element to wait on.  By modifying Wait to work like the other
primitives, we eliminate one reference to mElements.

We also tweak the MediaElement interface slightly to get a log name.