C++ nodes: Use BOOST_FOREACH to loop over mElements
We have a fair bit of code in Stage which loops over our flat list of
mElements.  Each of these callsites will need to be addressed in some
way before Halyard can support a hierachy of Node objects.

But since each of these loops uses iterators, the code is considerably
more verbose than necessary.  So this patch converts each 'for' loop
over mElements to use the BOOST_FOREACH macro.

Note that this macro involves some pretty hairy C++, and some folks on
StackOverflow believe that it carries a roughly 30% performance penalty:

  http://stackoverflow.com/questions/716865/c-boost-any-gotchas-with-boostforeach

Given the performance hit, I haven't decided whether or not we should
roll BOOST_FOREACH out to our entire codebase.  But I definitely want to
use it while working on Stage::mElements.