C++ nodes: Convert more mElements loops to use mNodes
Now that we've converted all our mElements loops to use BOOST_FOREACH,
we need to convert as many of them as possible to loop over mNodes.

In this patch, we only convert loops from using mElements to using
mNodes when we believe the order of iteration is irrelevant.  For
example, when calculating the intersection of several clipping regions,
we don't care about the order in which we process the individual
regions.

Note that the BOOST_FOREACH syntax isn't quite so nice when iterating
over a map's values--if we were feeling really inspired, we could build
a custom Boost.Range subclass that allowed us to iterate directly over
the values.  But that's a little too hairy for now.

There's still a couple more routines that we can convert over to mNodes
without too much work.