Node: Don't pass inStage to every constructor
We've been storing a pointer to the stage in every Node, which has
required us to add a parameter to every Node constructor.  This patch
removes the inStage/mStage/GetStage() machinery and just calls
wxGetApp().GetStage() whenever we need a pointer to the stage.

I'm cleaning this up now, because I'm about to add several new Node
subclasses, and don't want to keep typing this over and over again.