Elements pane: Add a primitive "Elements" pane
We modify the Node::RegisterWithParent and Node::UnregisterFromParent
methods to also register the Node with a new ElementsPane class.  For
now, this code is very simple, but it will serve as a starting point for
something more useful.

We also need to reverse the node deletion order in Stage::DeleteElements
so that each node is destroyed before its parent.  This requires
BOOST_REVERSE_FOREACH, which first became available in Boost 1.36.0.

Eyeballing element creation times on the Mac shows no significant
performance regressions (at least when the new palette is hidden).