Elements pane: Show the names of hidden nodes in grey
The actual changes to ElementsPane are fairly simple, but we also need
to clean up some rather messy overrides of Node::Show by factoring out
the subclass-specific logic into DoShow.

When reviewing this patch, pay particular attention to the semantics of
Widget::Show and Widget::DoShow, which _should_ be identical to the old
behavior.