C++ nodes: Composite using new Z-order code
This code extracts our compositing loop from Stage, and turns it into a
recursive operation on Node.  Note that we do not yet address
hit-testing or raising objects, nor do we address visibility.

We need to pay particular attention to the "drag layer" in this patch,
which we use to temporarily draw dragged objects on top of all other
elements.  This patch defines an element as being in the drag layer if
that element _or any of its ancestors_ (under the new Z-order semantics)
are in the drag layer.