C++ nodes: Hit-test using new Z-order code
This patch moves most of the hit-testing logic out of
Stage::FindLightWeightElement and into Node::FindNodeAt, and implements
hierarchical hit testing that matches the visual behavior of
RecursivelyCompositeInto.  (As in the old code, we ignore the base layer
vs. drag layer distinction when hit testing, which is dubious but
generally harmless--provided the drag layer is being used only for
objects that are being dragged-and-dropped or that are subclasses of
CursorElement.)

When reviewing this patch, please try to verify that we have preserved
the legacy behavior when HasLegacyZOrderAndVisibility returns true.
Here is a list of the changes made by this patch, with arguments as to
why these changes preserve the current behavior:

1) The old Stage::FindLightWeightElement function iterated over the
   nodes in creation order, but the new Node::FindNodeAt function
   iterates over the node hierarchy.

This should preserve the old behavior for nodes where
HasLegacyZOrderAndVisibility is true, because those nodes will all be
treated as children of mCurrentCard.  Their order doesn't _quite_ match
yet, because we haven't fixed RaiseToTop, but that should be the only
discrepancy.

2) The old FindLightWeightElement function iterated over mElements in
   forward order and took the last match.  The new FindNodeAt function
   iterates in reverse order and takes the first match.

3) The old FindLightWeightElement function could only return instances of
   Element where IsLightWeight returned true.  The new FindNodeAt function
   can theoretically return any node.

To verify this behavior, note that FindNodeAt can only return an element if
IsPointInNode returns true.  This should only happen for those classes
where IsLightWeight returned true, *plus* Card.  But Card can only be found
after checking all of its descendants, plus any group nodes for which
HasLegacyZOrderAndVisibility returns true.  And if the Card is returned, it
will be filtered out by the new FindLightWeightElementAt function.

OPEN QUESTION: Should we return the root node, and not the current card, as
the match of last resort?  Argument in favor: The card is currently
invisible, and so any mouse events should propagate to the root node.
Argument against: The card currently receives mouse events.  Any decision
we make should take into account the effects of someday adding child DCs.