C++ nodes: Send mouse-(enter|leave) to children of grabbed element
Under the new Z-order semantics, the children of a grabbed element will
typically be visible to the cursor-handling routines.  But this means
that when we move the mouse from the grabbed element and into one its
children, we receive a mouse-leave event with no corresponding
mouse-enter event.

This cause pressed buttons to be unpressed when we place the cursor over
the label.

To fix this, we slightly redefine the semantics of grabbing the cursor
when using the new Z-order semantics.  Specifically, we send mouse-enter
and mouse-leave events to the grabbed element and all its children.