C++ nodes: Hide and show elements using new Z-order semantics
When using the new Z-order and visibility semantics, hiding an element
will now automatically hide its descendants.  Showing that element again
will only show those descendants that were previously visible.

Here are some of the major changes:

 1) Quite a lot of visibility code is has moved out of Node and
    LightweightElement, and into Element.

 2) Most of the code that called Node::IsShown() now calls
    Node::IsVisible().  Cards and CardGroups are always visible, but
    Elements are only visible if (a) all their parents are visible and
    (b) the elements themselves are "shown".  This means that IsVisible
    is computed from the GetIsShown state of several nodes, but that
    GetIsShown is itself a purely element-local property.

 3) The DoShow function is now replaced by NotifyVisibilityChanged
    function.  This allows elements to show and hide associated wxWindow
    objects, Quake 2 overlays, etc., when the element's visibility
    changes.

When reviewing this patch, please pay particular attention to (a) the
correctness of RecursivelyCalculateVisibility, (b) what happens when you
show or hide an element where HasVisibleRepresentation() returns false,
and (c) the notifications that Node::Show used to send to the Stage and
the ElementsPane.