C++ nodes: Use current group member as legacy element parent
In order to simulate legacy Z-order semantics, we give every legacy
element a virtual parent of the current card; that way, they can be
interleaved with new style elements sensibly.  In between jumping to
cards, however, there are times when there is no current card, but the
elements still need to have a parent for purposes of Z-order and
visibility (for instance, when hiding or showing elements parented to
the group).

This patch modifies the legacy Z-order parent to be the current group
member, rather than the current card.  When we do this, we need to
remove the elements from the existing current group member, and add
them to the new one.  We have to do this when we're adding a new group
member node (which indicates that we're entering a child group or
card), or when we're deleting a group member node (leaving a group or
card, and thus its parent becoming the current group member).