Don't mark events as stale if their timestamp equals sMaxStaleTime
In order to avoid processing multiple clicks from the user when we do
something slow like creating a lot of elements in response to a click,
we mark events as stale if their timestamp is earlier than when the
previous event finished being processed.  We had been using a <=
comparison, but for some reason, in modern engines, many event
timestamps wound up being equal to the time when the previous event
finished processing, causing lots of valid events to be dropped.

Bug: http://iml.dartmouth.edu/fogbugz/default.php?12102