C++ nodes: Fix sementics of legacy children for current group member
When changing legacy elements to be artificial children of the current
group member rather than the current card, I had neglected to update
the logic for determining if an element should be treated as a child
when traversing downwards in the hierarchy.  This patch updates
GroupMember::IsChildForPurposeOfZOrderAndVisibility() to always return
true if that group member is the current group member, and otherwise
to fall back to the default behavior for Node.