Updater: Provide stronger guarantees for MANIFEST-DIFF
Our update downloader (written in Scheme) communicates with our update
installer (written in C++) in part through a file called MANIFEST-DIFF,
which contains the files that the downloader believes should be updated.
In anticipation of some new features that we need (renaming directories
which have changed only in case, and not downloading files that already
exist under a different name in the exisiting tree), this patch improves
the unit test for diff-manifests to ensure that we don't break anything
in the upcoming series of patches.