Updater: Only download files that we do not already have on disk
Because we have recently done some fairly major reorganizations of our
program layout, including our media and graphics, and we would like
to be able to update our program, we need to be careful not to download
again files that we already have.  We do this by filtering out from
our manifest diff not only the files we already have in the pool, but
also any files we find in any of our existing manifests (which should
match what we have on disk, unless the existing installation or a
previous update managed to break pretty badly somehow).

Note that this patch will cause the update installer to break if
launched from an update from this engine, because it expects to find
all files that it needs in the pool, while this patch makes the updater
assume that the update installer can find files in the existing tree.
A future patch in this series should be able to fix that, and may
be able to be rebased under this one so that this paragraph may be
removed.