UpdateInstaller: Add unit tests for parsing full manifest
We would like our UpdateInstaller.exe to figure out what needs to be copied
where directly from the original and update manifests, rather than from
the MANIFEST-DIFF provided by updater.ss.  This way, it can figure out
files that need to be deleted, do any case renames that it needs to,
and so on.

This patch adds some unit tests to ensure that the code used for parsing
MANIFEST-DIFF can also be used for parsing the whole manifests.