UpdateInstaller: Rename Manifest -> FileSpec, merge ManifestDiff
We have been using Manifest to represent more than just simple manifests;
we basically use it as a set of files, such as representing a manifest
diff.  This patch renames it to FileSet, and merges in ManifestDiff.
We also make all of the constructors on FileSet into static member
functions, to put them all on an equal footing and make it easier
to add more for future file set operations.