UpdateInstaller: Add support for copying files from tree to pool
Our updater needs to be able to find files in the tree that are missing
from the pool; we would like to avoid downloading files if we already
have them on disk, but they are copied or moved in the update.

This patch adds preliminary support for finding a file already in the
tree, and copying it into the pool.  This is a proof of concept that
passes the test case, rather than a final design; it could use substantial
refactoring and cleanup.

This patch also cleans up some of the inconsistently named APIs, to
use a Halyard style naming convention.