UpdateInstaller: Add notes on model behind new behavior
This file contains notes on how the whole updater update process (and
migration from Tamale file layout to Halyard file layout) is going to
work.  It includes notes on all the steps from the lazier downloader,
to the smarter update installer, to the two step update process, but
has particularly detailed notes on how the update installer can track
all of the changes it needs to make (moves and copies from both the
exising installed tree and the download pool).  It is a bit brief and
not yet quite complete, but it contains the work I've done so far in
trying to reason through this process.